creating empty DLL Encryptor. Some wiring.
