+++ build.sh
@@ -1,0 +1,13 @@
+cd /arrowdev/cpp/build
+
+cmake -GNinja \
+  -DARROW_WITH_SNAPPY=ON \
+  -DCMAKE_INSTALL_PREFIX=$(pwd)/install \
+  -DARROW_PARQUET=ON \
+  -DPARQUET_REQUIRE_ENCRYPTION=ON \
+  -DARROW_COMPUTE=ON \
+  -DARROW_PYTHON=ON \
+  -DCMAKE_BUILD_TYPE=Debug \
+  ..
+ninja install
+
